modularize service creation UI and move logic to useServiceForm hook
- Extracted form logic to `useServiceForm` hook for better separation of concerns.
- Split `Services` component into modular sections (`ServiceInfoSection`, `ServiceDetailsSection`, `CustomerInfoSection`, `CustomAttributesSection`, `AttachmentsSection`, `FormFooter`).
- Moved `FormDataType` and `CustomAttribute` types to `types/services.ts` for centralized type management.
- Removed `table.tsx` and renamed to `ServiceTable.tsx` for consistency.
- Reduced `Services` component from 454 to 33 lines, improving maintainability.
